app FSD-layer 형성
