+++ frontend/src/index.js
@@ -1,48 +1,12 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter, Route, Routes, useLocation } from 'react-router-dom';
+import ReactDOM from 'react-dom/client';
+import { BrowserRouter } from 'react-router-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import Main from './pages/HomePage/HomePage';
-import ArticleMain from './pages/ArticlePage/ArticleMain';
-import CodingMain from './pages/Coding-zone/CodingZoneMain';
-import EditPage from './pages/ArticlePage/Components/EditPage';
-import CodingClassRegist from './pages/Coding-zone/CodingZoneRegist';
-import CodingZoneAttendanceAssistant from './pages/Coding-zone/CodingZoneAttendanceAssistant';
-import CodingZoneMyAttendance from './pages/Coding-zone/CodingZoneMyAttendance';
-import CodingZoneAttendanceManager from './pages/Coding-zone/CodingZoneAttendanceManager';
-import NavBar from './widgets/layout/Header/Navbar';
-import Footer from './widgets/layout/Footer/Footer';
-import CreatePage from './pages/ArticlePage/Components/CreatePage';
-import ShowPage from './pages/ArticlePage/Components/ShowPage';
-import AuthHandle from './pages/AuthHandle';
- 
-
-const App = () => {
-  const location = useLocation();
-
-  return (
-    <>
-      <NavBar />
-      <Routes>
-        <Route path="/" element={<Main />} />
-        <Route path="/auth-handle" element={<AuthHandle />} /> 
-        <Route path="/article-main" element={<ArticleMain />} />
-        <Route path="/article-main/create" element={<CreatePage />} />
-        <Route path="/article-main/:articleNum/edit" element={<EditPage />} />
-        <Route path="/article-main/:articleNum" element={<ShowPage />} />
-        <Route path="/coding-zone" element={<CodingMain />} />
-        <Route path="/coding-zone/coding-class-regist" element={<CodingClassRegist />} />
-        <Route path="/coding-zone/Codingzone_Manager" element={<CodingZoneAttendanceAssistant />} />
-        <Route path="/coding-zone/Codingzone_Attendance" element={< CodingZoneMyAttendance/>} />
-        <Route path="/coding-zone/Codingzone_All_Attend" element={<CodingZoneAttendanceManager/>} />
-      </Routes>
-      {location.pathname !== '/' && <Footer />}
-    </>
-  );
-};
+import App from '../src/app/App';
 
 const container = document.getElementById('root');
 const root = ReactDOM.createRoot(container);
+
 root.render(
   <BrowserRouter>
     <App />
